+++ src/kernel/LinearByteKernel.cpp
@@ -33,7 +33,7 @@ void CLinearByteKernel::init_rescale()
 	for (long i=0; (i<lhs->get_num_vectors() && i<rhs->get_num_vectors()); i++)
 			sum+=compute(i, i);
 
-	if ( sum > (pow(2,8*sizeof(long))) )
+	if ( sum > (pow((double) 2, (double) 8*sizeof(long))) )
 		CIO::message("the sum %lf does not fit into integer of %d bits expect bogus results.\n", sum, 8*sizeof(long));
 	scale=sum/math.min(lhs->get_num_vectors(), rhs->get_num_vectors());
 }
